fix: Don't mask errors with diagnostic settings.
I've removed a layer of exception handling for the base diagnostic
settings filter for Graph-based resources, and the one exception we do
catch now, I've re-raised it after adding a custom error log message.

This is because it's not the norm to mask all errors in other filters,
and in this case, I think it best to let those errors propagate so
that we don't mistakenly produce incorrect results for a policy.