fix: Refactor auth-methods filter to work like other filters.
Instead of writing a filter to just annotate resources and require a
separate value filter to do actual filtering, redo the auth-methods
filter to be a value filter, with the values tied to the annotation
rather than the whole resource.  This is similar to other filters
across all cloud providers.